requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
pymongo==4.7.3
bcrypt==4.1.3
marshmallow==3.21.3
python-dotenv==1.0.1
cloudinary==1.40.0
flask-limiter==3.7.0
flasgger==0.9.7.1
dnspython==2.6.1
